+++ src/AnimeFeedManager.Storage/Domain/SeasonStorage.cs
@@ -5,6 +5,6 @@ namespace AnimeFeedManager.Storage.Domain
     public class SeasonStorage: TableEntity
     {
         public string? Season { get; set; }
-        public ushort Year { get; set; }
+        public int Year { get; set; } // Azure tables only works with Int and Int64
     }
 }
